Move JabRefExecutorService to logic as HeadlessExecutorService
