+++ src/main/java/org/jabref/gui/journals/AbbreviateAction.java
@@ -11,7 +11,6 @@ import java.util.stream.Collectors;
 import javax.swing.undo.UndoManager;
 
 import org.jabref.gui.DialogService;
-import org.jabref.gui.JabRefExecutorService;
 import org.jabref.gui.LibraryTab;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.actions.ActionHelper;
@@ -23,6 +22,7 @@ import org.jabref.gui.util.TaskExecutor;
 import org.jabref.logic.journals.JournalAbbreviationPreferences;
 import org.jabref.logic.journals.JournalAbbreviationRepository;
 import org.jabref.logic.l10n.Localization;
+import org.jabref.logic.util.HeadlessExecutorService;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.field.FieldFactory;
@@ -111,7 +111,7 @@ public class AbbreviateAction extends SimpleCommand {
                 .collect(Collectors.toSet());
 
         // Execute the callables and wait for the results.
-        List<Future<Boolean>> futures = JabRefExecutorService.INSTANCE.executeAll(tasks);
+        List<Future<Boolean>> futures = HeadlessExecutorService.INSTANCE.executeAll(tasks);
 
         // Evaluate the results of the callables.
         long count = futures.stream().filter(future -> {
