+++ src/main/java/org/jabref/gui/push/PushToLyx.java
@@ -8,10 +8,10 @@ import java.nio.charset.StandardCharsets;
 import java.util.List;
 
 import org.jabref.gui.DialogService;
-import org.jabref.gui.JabRefExecutorService;
 import org.jabref.gui.icon.IconTheme;
 import org.jabref.gui.icon.JabRefIcon;
 import org.jabref.logic.l10n.Localization;
+import org.jabref.logic.util.HeadlessExecutorService;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.preferences.PreferencesService;
@@ -86,7 +86,7 @@ public class PushToLyx extends AbstractPushToApplication {
 
         final File lyxpipe = lp;
 
-        JabRefExecutorService.INSTANCE.executeAndWait(() -> {
+        HeadlessExecutorService.INSTANCE.executeAndWait(() -> {
             try (FileWriter fw = new FileWriter(lyxpipe, StandardCharsets.UTF_8); BufferedWriter lyxOut = new BufferedWriter(fw)) {
                 String citeStr = "LYXCMD:sampleclient:citation-insert:" + keyString;
                 lyxOut.write(citeStr + "\n");
